+++ package/naruse-h5/components/index.ts
@@ -11,6 +11,8 @@ import Textarea from './textarea/index'
 import { getNaruseConfig } from "../core/init";
 import { functionalizae } from 'core/Component';
 import WebView from './web-view';
+import Radio from './radio';
+import RadioGroup from './radio-group';
 
 /** 组件映射表 */
 const componentReflectMap = {
@@ -23,6 +25,8 @@ const componentReflectMap = {
     'scroll-view': ScrollView,
     textarea: Textarea,
     'web-view': WebView,
+    'radio': Radio,
+    'radio-group': RadioGroup
 }
 
 /**
@@ -57,7 +61,7 @@ const naruseCreateElement = (type: string, props: any, ...children: string[]) =>
 };
 
 
-const rpxReg = /(\d+)\s?rpx/g;
+const rpxReg = /([\d.]+)\s?rpx/g;
 
 const parsePx = val => {
     if (typeof val !== 'string') return val;