+++ package/fs-aw/CHANGELOG.md
@@ -1,5 +1,17 @@
 # fs-aw
 
+## 0.1.5
+
+### Patch Changes
+
+- #FEAT: H5 新增 radio-group、radio
+
+## 0.1.4
+
+### Patch Changes
+
+- #FIX: rpx\px 有小数时转换错误
+
 ## 0.1.3
 
 ### Patch Changes