+++ package/naruse-h5/CHANGELOG.md
@@ -1,5 +1,17 @@
 # naruse-h5
 
+## 0.7.5
+
+### Patch Changes
+
+- #FEAT: H5 新增 radio-group、radio
+
+## 0.7.4
+
+### Patch Changes
+
+- #FIX: rpx\px 有小数时转换错误
+
 ## 0.7.3
 
 ### Patch Changes