+++ package/naruse-parser/CHANGELOG.md
@@ -1,5 +1,17 @@
 # naruse-parser
 
+## 0.2.5
+
+### Patch Changes
+
+- #FEAT: H5 新增 radio-group、radio
+
+## 0.2.4
+
+### Patch Changes
+
+- #FIX: rpx\px 有小数时转换错误
+
 ## 0.2.3
 
 ### Patch Changes