+++ package/naruse-plugin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # naruse-plugin
 
+## 1.3.5
+
+### Patch Changes
+
+- #FEAT: H5 新增 radio-group、radio
+
+## 1.3.4
+
+### Patch Changes
+
+- #FIX: rpx\px 有小数时转换错误
+
 ## 1.3.3
 
 ### Patch Changes